Update Node.swift (#3)
`xmlGetLineNo` is defined as a function returning a `Long`.  The correct type to map that to Swift is `CLong`.  Due to the differences in LP64 and LLP64, `long` is not mapped equivalently on all environments.  However, to maintain the API, use a `numericCast` to convert the integral type from any compatible type from the C interface.